pyproject.toml
[project]
name = "ms-graph-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Brian Cusack", email = "briancusack78@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "msgraph-sdk>=1.29.0",
]

[project.scripts]
ms-graph-test = "ms_graph_test.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
